requirements.txt
aio-celery==0.20.0
aio-pika==9.5.4
aiofiles==24.1.0
aiormq==6.8.1
aitertools==0.1.0
alembic==1.13.1
amqp==5.3.1
annotated-types==0.7.0
anyio==4.8.0
async-timeout==5.0.1
asyncpg==0.29.0
bcrypt==4.2.1
billiard==4.2.1
black==24.10.0
celery==5.4.0
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cryptography==44.0.0
databases==0.9.0
distro==1.9.0
dnspython==2.7.0
ecdsa==0.19.0
email_validator==2.2.0
ephem==4.1.6
exceptiongroup==1.2.2
fastapi==0.109.0
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
Jinja2==3.1.5
kombu==5.4.2
Mako==1.3.8
MarkupSafe==3.0.2
multidict==6.1.0
mypy-extensions==1.0.0
openai==1.12.0
packaging==24.2
pamqp==3.3.0
passlib==1.7.4
pathspec==0.12.1
pillow==11.1.0
platformdirs==4.3.6
prompt_toolkit==3.0.48
propcache==0.2.1
pyasn1==0.6.1
pycparser==2.22
pydantic==2.10.5
pydantic-settings==2.1.0
pydantic_core==2.27.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-jose==3.3.0
python-multipart==0.0.6
redis==5.2.1
requests==2.32.3
rsa==4.9
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.25
sqlalchemy_celery_beat==0.8.0
sse-starlette==2.0.0
starlette==0.35.1
tqdm==4.67.1
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.3.0
uvicorn==0.27.0
uvloop==0.21.0
vine==5.1.0
wcwidth==0.2.13
yarl==1.18.3
